+++ TagsCloudVisualization/TagsCloudVisualization/App/Cloud/Image/IVisualizer.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using System.Drawing;
+
+namespace TagsCloudVisualization
+{
+    public interface IVisualizer
+    {
+        Image Render(IEnumerable<GraphicWord> words, int width, int height, IWordPalette palette);
+    }
+}
