+++ TagsCloudVisualization/TagsCloudVisualization/App/Cloud/Image/ImageSettings.cs
@@ -1,0 +1,10 @@
+﻿using System.Drawing;
+
+namespace TagsCloudVisualization
+{
+    public class ImageSettings
+    {
+        public Point Center { get; set; }
+        public Size Size { get; set; }
+    }
+}
