+++ TagsCloudVisualization/TagsCloudVisualization/App/Cloud/Layouters/ICloudLayouter.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using System.Drawing;
+
+namespace TagsCloudVisualization
+{
+    public interface ICloudLayouter
+    {
+        void Process(IEnumerable<GraphicWord> graphicWord, ISizeDefiner sizeDefiner, Point center);
+    }
+}
