+++ TagsCloudVisualization/TagsCloudVisualization/App/Cloud/Words/ISizeDefiner.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace TagsCloudVisualization
+{
+    public interface ISizeDefiner
+    {
+        Size GetSize(GraphicWord graphicWord);
+    }
+}
