+++ TagsCloudVisualization/TagsCloudVisualization/App/Cloud/Words/NWordSizer.cs
@@ -1,0 +1,13 @@
+﻿using System.Drawing;
+using System.Windows.Forms;
+
+namespace TagsCloudVisualization
+{
+    public class NWordSizer : ISizeDefiner
+    {
+        public Size GetSize(GraphicWord graphicWord)
+        {
+            return TextRenderer.MeasureText(graphicWord.Value, graphicWord.Font);
+        }
+    }
+}
