+++ TagsCloudVisualization/TagsCloudVisualization/App/File workers/IFileReader.cs
@@ -1,0 +1,7 @@
+﻿namespace TagsCloudVisualization
+{
+    public interface IFileReader
+    {
+        string Read(string fileName);
+    }
+}
