+++ TagsCloudVisualization/TagsCloudVisualization/App/File workers/ImageSaver.cs
@@ -1,0 +1,12 @@
+﻿using System.Drawing;
+
+namespace TagsCloudVisualization
+{
+    public static class ImageSaver
+    {
+        public static void WriteToFile(string fileName, Image bitmap)
+        {
+            bitmap.Save(fileName, bitmap.RawFormat);
+        }
+    }
+}
