+++ TagsCloudVisualization/TagsCloudVisualization/App/File workers/TxtReader.cs
@@ -1,0 +1,12 @@
+﻿using System.IO;
+
+namespace TagsCloudVisualization
+{
+    public class TxtReader : IFileReader
+    {
+        public string Read(string fileName)
+        {
+            return File.ReadAllText(fileName);
+        }
+    }
+}
