--- TagsCloudVisualization/TagsCloudVisualization/App/Infrastructure/Cloud/ICloudLayouter.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Drawing;
-
-namespace TagsCloudVisualization
-{
-    public interface ICloudLayouter
-    {
-        void Process(IEnumerable<GraphicWord> graphicWord, ISizeDefiner sizeDefiner, Point center);
-    }
-}
