--- TagsCloudVisualization/TagsCloudVisualization/App/Infrastructure/Cloud/IFileReader.cs
@@ -1,7 +1,0 @@
-﻿namespace TagsCloudVisualization
-{
-    public interface IFileReader
-    {
-        string Read(string fileName);
-    }
-}
