--- TagsCloudVisualization/TagsCloudVisualization/App/Infrastructure/Cloud/ISizeDefiner.cs
@@ -1,9 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagsCloudVisualization
-{
-    public interface ISizeDefiner
-    {
-        Size GetSize(GraphicWord graphicWord);
-    }
-}
