--- TagsCloudVisualization/TagsCloudVisualization/App/Infrastructure/Cloud/ImageSettings.cs
@@ -1,10 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagsCloudVisualization
-{
-    public class ImageSettings
-    {
-        public Point Center { get; set; }
-        public Size Size { get; set; }
-    }
-}
