--- TagsCloudVisualization/TagsCloudVisualization/App/Infrastructure/Cloud/NWordSizer.cs
@@ -1,13 +1,0 @@
-﻿using System.Drawing;
-using System.Windows.Forms;
-
-namespace TagsCloudVisualization
-{
-    public class NWordSizer : ISizeDefiner
-    {
-        public Size GetSize(GraphicWord graphicWord)
-        {
-            return TextRenderer.MeasureText(graphicWord.Value, graphicWord.Font);
-        }
-    }
-}
