--- TagsCloudVisualization/TagsCloudVisualization/App/Infrastructure/SettingsForm.cs
@@ -1,24 +1,0 @@
-﻿using System.Windows.Forms;
-
-namespace TagsCloudVisualization
-{
-    public class SettingsForm<TSettings> : Form
-    {
-        public SettingsForm(TSettings settings)
-        {
-            var okButton = new Button
-            {
-                Text = "OK",
-                DialogResult = DialogResult.OK,
-                Dock = DockStyle.Bottom,
-            };
-            Controls.Add(okButton);
-            Controls.Add(new PropertyGrid
-            {
-                SelectedObject = settings,
-                Dock = DockStyle.Fill
-            });
-            AcceptButton = okButton;
-        }
-    }
-}
