--- TagsCloudVisualization/TagsCloudVisualization/App/TxtReader.cs
@@ -1,12 +1,0 @@
-﻿using System.IO;
-
-namespace TagsCloudVisualization
-{
-    public class TxtReader : IFileReader
-    {
-        public string Read(string fileName)
-        {
-            return File.ReadAllText(fileName);
-        }
-    }
-}
