+++ TagsCloudVisualization/TagsCloudVisualization/Stopwords.txt
@@ -1,0 +1,39 @@
+﻿Я
+мы
+ты
+вы
+он
+она
+оно
+они
+мой
+твой
+наш
+ваш
+свой
+и
+мои
+моего
+моему
+моей
+моим
+кто
+что
+какой
+чей
+как
+где
+куда
+когда
+а
+у
+в
+на
+за
+под
+к
+около
+перед
+ему
+это
+этот
